🏗️ Move components related to members list feature into new dedicated folder
