+++ src/features/admin/members-list/components/AdminMembers/style.scss
@@ -1,4 +1,4 @@
-@use '../../../styles/settings/breakpoints' as b;
+@use '../../../../../styles/settings/breakpoints' as b;
 
 .c-admin-members {
     &__title {
