+++ src/features/admin/members-list/components/MemberCard/style.scss
@@ -1,5 +1,5 @@
-@use '../../../styles/settings/breakpoints' as b;
-@use '../../../styles/settings/colors' as colors;
+@use '../../../../../styles/settings/breakpoints' as b;
+@use '../../../../../styles/settings/colors' as colors;
 
 .c-member-card {
     &__group {
