+++ src/pages/Administration/index.jsx
@@ -1,5 +1,5 @@
 import { AuthenticatedLayout } from '../../layouts/AuthenticatedLayout'
-import { AdminMembers } from '../../components/Admin/AdminMembers'
+import { AdminMembers } from '../../features/admin/members-list/components/AdminMembers'
 import { ScrollTopButton } from '../../components/Buttons/ScrollTopButton'
 import './style.scss'
